+++ code-execution-engines/langchain4j-code-execution-engine-judge0/src/main/java/dev/langchain4j/code/judge0/Judge0JavaScriptEngine.java
@@ -1,5 +1,6 @@
-package dev.langchain4j.code;
+package dev.langchain4j.code.judge0;
 
+import dev.langchain4j.code.CodeExecutionEngine;
 import dev.langchain4j.internal.Json;
 import okhttp3.*;
 import org.slf4j.Logger;
