+++ code-execution-engines/langchain4j-code-execution-engine-judge0/src/test/java/dev/langchain4j/code/judge0/JavaScriptCodeFixerTest.java
@@ -1,4 +1,4 @@
-package dev.langchain4j.code;
+package dev.langchain4j.code.judge0;
 
 import org.junit.jupiter.api.Test;
